backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
sentence-transformers
faiss-cpu
pytesseract
pdfplumber
python-docs
PyMupdf
python-multipart
aiofiles
Pillow
groq